# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Application REST-API scheme",
        "title": "Golang Effective Mobile API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
        "/subscriptions": {
            "get": {
                "description": "Просмотр списка всех подписок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Список подписок",
                "responses": {
                    "200": {
                        "description": "Список подписок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_transport_http.SubscriptionDTOResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать новую подписку в системе.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Создать подписку",
                "parameters": [
                    {
                        "description": "CreateSubscription тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_transport_http.SubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданная подписка",
                        "schema": {
                            "$ref": "#/definitions/internal_features_transport_http.CreateSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/total": {
            "get": {
                "description": "Получение общей суммы подписок за указанный период с опциональной фильтрацией по id пользователя и названию сервиса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получение общей суммы подписок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала периода в формате MM-YYYY",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания периода в формате MM-YYYY",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Фильтрация подписок по ID пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтрация подписок по названию сервиса",
                        "name": "service_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Общая стоимость подписок",
                        "schema": {
                            "$ref": "#/definitions/internal_features_transport_http.TotalPriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "description": "Получение конкретной подписки по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получение подписки",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID получаемой подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "подписка успешно найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_transport_http.GetSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию об уже существующей в системе подписке.\nТело запроса должно содержать все поля подписки (как при создании).\nПоле `end_date` опционально; если оно не передано или равно null, дата окончания сбрасывается.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Обновить запись о подписке полностью",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID обновляемой подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PutSubscription тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_transport_http.SubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно измененная подписка",
                        "schema": {
                            "$ref": "#/definitions/internal_features_transport_http.PutSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить существующую в системе подписку по ее ID",
                "tags": [
                    "subscriptions"
                ],
                "summary": "Удаление подписки",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID удаляемой подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление подписки"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Subscription not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Vadick-do_Effective-Mobile_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
        "internal_features_transport_http.CreateSubscriptionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "end_date": {
                    "type": "string",
                    "example": "08-2025"
                },
                "id": {
                    "type": "string",
                    "example": "a3f2e8d1-9b4c-4c7f-b2d6-1e5a8c9f0b3d"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-06-24T16:15:18.62982Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_features_transport_http.GetSubscriptionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "end_date": {
                    "type": "string",
                    "example": "08-2025"
                },
                "id": {
                    "type": "string",
                    "example": "a3f2e8d1-9b4c-4c7f-b2d6-1e5a8c9f0b3d"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-06-24T16:15:18.62982Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_features_transport_http.PutSubscriptionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "end_date": {
                    "type": "string",
                    "example": "08-2025"
                },
                "id": {
                    "type": "string",
                    "example": "a3f2e8d1-9b4c-4c7f-b2d6-1e5a8c9f0b3d"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-06-24T16:15:18.62982Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_features_transport_http.SubscriptionDTOResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "end_date": {
                    "type": "string",
                    "example": "08-2025"
                },
                "id": {
                    "type": "string",
                    "example": "a3f2e8d1-9b4c-4c7f-b2d6-1e5a8c9f0b3d"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-06-24T16:15:18.62982Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_features_transport_http.SubscriptionRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "08-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_features_transport_http.TotalPriceResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer",
                    "example": 1000
                }
            }
        }
    }
}